requirements.txt
atari-py==0.1.1
bleach==2.1.1
boto==2.48.0
boto3==1.4.8
botocore==1.8.8
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
future==0.16.0
gensim==3.1.0
graphviz==0.8.1
gym==0.9.4
h5py==2.7.1
html5lib==0.9999999
idna==2.6
invoke==0.22.1
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.30.2
jupyterlab-launcher==0.6.0
Keras==2.1.2
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
notebook==5.2.2
numpy==1.13.3
opencv-python==3.3.0.10
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
pydot==1.2.4
pydot-ng==1.0.0
pydotplus==2.0.2
pyglet==1.3.0
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
s3transfer==0.1.12
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.5.4
tabulate==0.8.2
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
tqdm==4.19.6
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.8
